Move some functions to /lib
