+++ http-practice/lib/accepts.js
@@ -1,0 +1,7 @@
+module.exports = function accepts(req) {
+  if (req.headers.accept.indexOf("application/json") !== -1) {
+    return "json";
+  } else {
+    return "html";
+  }
+}
