+++ http-practice/lib/generateNum.js
